Updated pathways, removed BIT pathway. Seperate code into new file that: adds ele/ecs modules to modules table. Seperate code into new file that links all pathways with their correct modules.
